# jrag-model/src/main/resources/openapi-interface.yaml
openapi: 3.0.3
info:
  title: ai operate - OpenAPI 3.0
  description: |-
    This is ai operate OpenAPI.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:30110
    description: 本机调试环境
...
paths:
...
  /v1/rest/jrag/users:
    get:
      tags: [ User ]
      summary: 获取用户列表
      description: 获取用户列表
      operationId: getUserList
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './openapi-model.yaml#/components/schemas/UserListDto'
    post:
      tags: [ User ]
      summary: 创建用户
      description: 创建用户
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './openapi-model.yaml#/components/schemas/UserCreateRequestDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './openapi-model.yaml#/components/schemas/UserDto'
    delete:
      tags: [ User ]
      summary: 删除用户
      description: 删除用户
      operationId: deleteUser
      parameters:
        - in: query
          name: user-id
          schema:
            type: string
          required: true
          description: 用户ID
      responses:
        '200':
          description: OK
  /v1/rest/jrag/users/role:
    put:
      tags: [ User ]
      summary: 修改用户角色
      description: 修改用户角色
      operationId: updateUserRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './openapi-model.yaml#/components/schemas/UserRoleUpdateRequestDto'
      responses:
        '200':
          description: OK
  /v1/rest/jrag/users/password:
    put:
      tags: [ User ]
      summary: 修改用户密码
      description: 修改用户密码
      operationId: updateUserPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './openapi-model.yaml#/components/schemas/UserPasswordUpdateRequestDto'
      responses:
        '200':
          description: OK
...
  /v1/rest/jrag/property/get:
    post:
      tags: [ Property ]
      summary: 获取属性
      description: 获取属性
      operationId: getProperty
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
  /v1/rest/jrag/property:
    put:
      tags: [ Property ]
      summary: 覆盖属性
      description: 覆盖属性
      operationId: putProperty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: './openapi-model.yaml#/components/schemas/PropertyDto'
      responses:
        '200':
          description: OK
  /v1/rest/jrag/mcp/config:
    get:
      tags: [ MCP ]
      summary: 获取 MCP 配置
      description: 获取 MCP 配置
      operationId: getMcpConfig
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
    put:
      tags: [ MCP ]
      summary: 保存 MCP 配置
      description: 保存 MCP 配置
      operationId: updateMcpConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...